refactor: extract buildDipoleWires to antennaGeometry
