+++ src/store/antennaGeometry.ts
@@ -2,6 +2,7 @@ import { cleanZero } from "../utils/math";
 import {
   SLOPING_V_MIN_TIP_Z_M,
   wavelengthMeters,
+  MAIN_WIRE_TAG,
   LEFT_LEG_TAG,
   RIGHT_LEG_TAG,
   FEED_BRIDGE_TAG,
@@ -1101,4 +1102,83 @@ export function buildFoldedAntennaWires(params: FoldedAntennaWiresParams): Wire[
     // Right end connector across the aperture.
     createWire(pts.rightFed, pts.rightOpp, segs.connSegs, FOLDED_DIPOLE_CONNECTOR_TAG),
   ];
+}
+
+export interface DipoleWiresParams {
+  length: number;
+  height: number;
+  orientation: Orientation;
+  wireRadius: number;
+  segments: number;
+  layout: { offset: number; shield: { bottomZ: number; radius: number } | null } | null;
+}
+
+export function buildDipoleWires(params: DipoleWiresParams): Wire[] {
+  const { length, height: h, orientation, wireRadius, segments, layout } = params;
+  const half = length / 2;
+  const [dx, dy] = orientationVector(orientation);
+
+  if (!layout) {
+    return [{
+      start: [cleanZero(-half * dx), cleanZero(-half * dy), h],
+      end: [cleanZero(half * dx), cleanZero(half * dy), h],
+      radius: wireRadius,
+      segments: segments,
+      tag: MAIN_WIRE_TAG,
+    }];
+  }
+
+  const offset = layout.offset;
+  const bridgeHalf = FEED_BRIDGE_LENGTH_M / 2;
+
+  const bridgeStart: [number, number, number] = [
+    cleanZero((offset - bridgeHalf) * dx),
+    cleanZero((offset - bridgeHalf) * dy),
+    h,
+  ];
+  const bridgeEnd: [number, number, number] = [
+    cleanZero((offset + bridgeHalf) * dx),
+    cleanZero((offset + bridgeHalf) * dy),
+    h,
+  ];
+
+  const leftTip: [number, number, number] = [cleanZero(-half * dx), cleanZero(-half * dy), h];
+  const rightTip: [number, number, number] = [cleanZero(half * dx), cleanZero(half * dy), h];
+
+  const dist = (p1: [number, number, number], p2: [number, number, number]) =>
+    Math.sqrt((p1[0] - p2[0]) ** 2 + (p1[1] - p2[1]) ** 2 + (p1[2] - p2[2]) ** 2);
+
+  const leftLen = dist(leftTip, bridgeStart);
+  const rightLen = dist(rightTip, bridgeEnd);
+
+  const totalSeg = segments;
+  const segDensity = totalSeg / length;
+  const leftSeg = Math.max(3, oddRound(leftLen * segDensity));
+  const rightSeg = Math.max(3, oddRound(rightLen * segDensity));
+
+  return [
+    {
+      start: leftTip, end: bridgeStart,
+      radius: wireRadius,
+      segments: leftSeg,
+      tag: LEFT_LEG_TAG,
+    },
+    {
+      start: bridgeEnd, end: rightTip,
+      radius: wireRadius,
+      segments: rightSeg,
+      tag: RIGHT_LEG_TAG,
+    },
+    {
+      start: bridgeStart, end: bridgeEnd,
+      radius: wireRadius,
+      segments: 1,
+      tag: FEED_BRIDGE_TAG,
+    },
+  ];
+}
+
+export function oddRound(v: number): number {
+  const n = Math.max(1, Math.round(v));
+  return n % 2 === 0 ? n + 1 : n;
 }
